+++ lib/include/sparseUtils.cc
@@ -8,7 +8,7 @@
  Utility to count how many entries above a specified value exist in a float array
  ---------------------------------------------------------------------*/
 template <class DATATYPE>
-unsigned int countEntriesAbove(DATATYPE * Array, int sz, float includeAbove)
+unsigned int countEntriesAbove(DATATYPE * Array, int sz, double includeAbove)
 {
 	int count = 0;
 	for (int i = 0; i < sz; i++) {
